+++ app/lib/utils/helper.ts
@@ -1,0 +1,12 @@
+export function darkenColor(hex: string, percent: number): string {
+    const num = parseInt(hex.slice(1), 16);
+    let r = (num >> 16) - Math.round((num >> 16) * percent / 100);
+    let g = ((num >> 8) & 0x00FF) - Math.round(((num >> 8) & 0x00FF) * percent / 100);
+    let b = (num & 0x0000FF) - Math.round((num & 0x0000FF) * percent / 100);
+  
+    r = Math.max(0, r);
+    g = Math.max(0, g);
+    b = Math.max(0, b);
+  
+    return `#${(r << 16 | g << 8 | b).toString(16).padStart(6, '0')}`;
+}
