+++ app/lib/utils/index.ts
@@ -1,0 +1,2 @@
+import { darkenColor } from './helper';
+export { darkenColor };
